Move filesystem.hpp to src
