+++ src/plugins/codemon/codemon.h
@@ -109,7 +109,7 @@
 #include <memory>
 #include <set>
 #include <glib.h>
-#include <plugins/filesystem.hpp>
+#include "filesystem.hpp"
 #include "plugins/private.h"
 #include "plugins/plugins_ex.h"
