+++ src/plugins/ipt/ipt.cpp
@@ -110,7 +110,7 @@
 #include <libvmi/libvmi.h>
 
 #include "ipt.h"
-#include <plugins/filesystem.hpp>
+#include "filesystem.hpp"
 #include "plugins/output_format.h"
 #include "private.h"
